backend/requirements.txt
fastapi
uvicorn
sqlalchemy
SQLAlchemy-Utils
alembic
asyncpg
psycopg2-binary
insightface
onnxruntime
python-multipart
pydantic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
opencv-python
numpy
gunicorn
pytest
httpx
